adding rename and filtering options for columns in output
