+++ topiary/__init__.py
@@ -21,7 +21,7 @@ from .sequence_helpers import (
 )
 from . import commandline_args
 
-__version__ = '0.0.19'
+__version__ = '0.0.20'
 
 __all__ = [
     "LazyLigandomeDict",
