+++ topiary/commandline_args/errors.py
@@ -1,0 +1,33 @@
+
+# Copyright (c) 2016. Mount Sinai School of Medicine
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+"""
+Commandline arguments related to error handling
+"""
+
+from __future__ import print_function, division, absolute_import
+
+def add_error_args(arg_parser):
+    error_group = arg_parser.add_argument_group(
+        title="Errors",
+        description="Options for error handling")
+
+    error_group.add_argument(
+        "--skip-variant-errors",
+        default=False,
+        action="store_true",
+        help="Skip variants which cause runtime errors of any kind")
+
+    return error_group
